requirements.txt
fastapi
uvicorn
rembg
onnxruntime
python-multipart
pillow
python-dotenv
